+++ maeumgagym-infrastructure/src/main/kotlin/com/info/maeumgagym/response/writer/ErrorLogResponseWriter.kt
@@ -1,13 +1,68 @@
 package com.info.maeumgagym.response.writer
 
 import com.info.maeumgagym.error.log.ErrorLog
+import java.time.LocalDateTime
 import javax.servlet.http.HttpServletResponse
 
+/**
+ * [ErrorLog] 및 [Exception]을 [HttpServletResponse]로 작성하기 위한 [ResponseWriter].
+ *
+ * [DefaultResponseWriter]를 내부 변수로 갖고 있어, [setBody]와 [setDefaultSetting]는 [DefaultResponseWriter]를 통해 Proxt 형태로 구현됨.
+ *
+ * @author Daybreak312
+ * @since 12-03-2024
+ */
 abstract class ErrorLogResponseWriter : ResponseWriter {
 
+    /**
+     * [DefaultResponseWriter.setBody]와 같은 동작을 하지만, 사용하지 않을 것을 권고.
+     *
+     * @author Daybreak312
+     * @since 12-03-2024
+     */
+    abstract override fun setBody(response: HttpServletResponse, `object`: Any): HttpServletResponse
+
+    /**
+     * [DefaultResponseWriter.setDefaultSetting]와 같은 동작을 하지만, 사용하지 않을 것을 권고.
+     *
+     * @author Daybreak312
+     * @since 12-03-2024
+     */
+    abstract override fun setDefaultSetting(response: HttpServletResponse, status: Int): HttpServletResponse
+
+    /**
+     * 인자로 받은 [ErrorLog]와 [Exception]의 정보를 기반으로 [HttpServletResponse]를 작성.
+     *
+     * 사용되는 정보는 [ErrorLogResponse]의 Docs 참고.
+     *
+     * @see [ErrorLogResponse]
+     *
+     * @author Daybreak312
+     * @since 12-03-2024
+     */
     abstract fun writeResponseWithErrorLogAndException(
         response: HttpServletResponse,
         errorLog: ErrorLog,
         e: Exception
     ): HttpServletResponse
 }
+
+/**
+ * Error의 정보를 담고 있는 Response
+ *
+ * @param status 상태 코드.
+ * @param message 예외 메세지.
+ * @param errorLogId 저장된 에러 로그의 Key.
+ * @param timestamp timestamp
+ * @param map Field Error 등 경우에 따라 Key-Value의 정보가 추가적으로 필요할 경우 사용
+ *
+ * @author Daybreak312
+ * @since 12-03-2024
+ */
+data class ErrorLogResponse(
+    val status: Int,
+    val message: String?,
+    val errorLogId: String,
+    val timestamp: LocalDateTime = LocalDateTime.now(),
+    val map: Map<String, String> = mapOf()
+)
