+++ maeumgagym-infrastructure/src/main/kotlin/com/info/maeumgagym/response/writer/impl/DefaultResponseWriterImpl.kt
@@ -7,6 +7,12 @@ import org.springframework.stereotype.Component
 import java.nio.charset.StandardCharsets
 import javax.servlet.http.HttpServletResponse
 
+/**
+ * Docs는 상위 추상 클래스에 존재
+ *
+ * @author Daybreak312
+ * @since 12-03-2024
+ */
 @Component
 private class DefaultResponseWriterImpl(
     private val objectMapper: ObjectMapper
