+++ maeumgagym-infrastructure/src/main/kotlin/com/info/maeumgagym/response/writer/impl/ErrorLogResponseWriterImpl.kt
@@ -3,11 +3,17 @@ package com.info.maeumgagym.response.writer.impl
 import com.info.maeumgagym.common.exception.PresentationValidationException
 import com.info.maeumgagym.error.log.ErrorLog
 import com.info.maeumgagym.response.writer.DefaultResponseWriter
+import com.info.maeumgagym.response.writer.ErrorLogResponse
 import com.info.maeumgagym.response.writer.ErrorLogResponseWriter
 import org.springframework.stereotype.Component
-import java.time.LocalDateTime
 import javax.servlet.http.HttpServletResponse
 
+/**
+ * Docs는 상위 추상 클래스에 존재
+ *
+ * @author Daybreak312
+ * @since 12-03-2024
+ */
 @Component
 private class ErrorLogResponseWriterImpl(
     private val defaultResponseWriter: DefaultResponseWriter
@@ -39,20 +45,3 @@ private class ErrorLogResponseWriterImpl(
         )
     }
 }
-
-/**
- * Error의 정보를 담고 있는 Response
- *
- * @param status 상태 코드.
- * @param message 예외 메세지.
- * @param errorLogId 저장된 에러 로그의 Key.
- * @param timestamp timestamp
- * @param map Field Error 등 경우에 따라 Key-Value의 정보가 추가적으로 필요할 경우 사용
- */
-data class ErrorLogResponse(
-    val status: Int,
-    val message: String?,
-    val errorLogId: String,
-    val timestamp: LocalDateTime = LocalDateTime.now(),
-    val map: Map<String, String> = mapOf()
-)
